Insert mode panes for trigger, side panel, and context menu
