+++ src/devTools/editor/hooks/useRemove.ts
@@ -15,58 +15,19 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IExtension } from "@/core";
 import { actions, FormState } from "@/devTools/editor/editorSlice";
-import { useDispatch, useSelector } from "react-redux";
-import { useCallback, useContext, useMemo } from "react";
-import { extensionToFormState } from "@/devTools/editor/extensionPoints/adapter";
-import { reportError } from "@/telemetry/logging";
+import { useCallback, useContext } from "react";
 import { DevToolsContext } from "@/devTools/context";
 import { useToasts } from "react-toast-notifications";
 import { useFormikContext } from "formik";
+import { useDispatch } from "react-redux";
+import { useModals } from "@/components/ConfirmationModal";
 import { uninstallContextMenu } from "@/background/devtools";
 import * as nativeOperations from "@/background/devtools";
 import { optionsSlice } from "@/options/slices";
-import { useModals } from "@/components/ConfirmationModal";
-import { useAsyncState } from "@/hooks/common";
-import axios from "axios";
-import { EditablePackage } from "@/devTools/editor/hooks/useCreate";
-import { makeURL } from "@/hooks/fetch";
-import { getExtensionToken } from "@/auth/token";
-import { RootState } from "@/devTools/store";
+import { reportError } from "@/telemetry/logging";
 
-const selectEditor = (x: RootState) => x.editor;
-
-export function useReset(
-  installed: IExtension[],
-  element: FormState
-): () => void {
-  const dispatch = useDispatch();
-  const { showConfirmation } = useModals();
-
-  return useCallback(async () => {
-    const confirm = await showConfirmation({
-      title: "Reset Brick?",
-      message: "Any changes you made since the last save will be lost",
-      submitCaption: "Reset",
-    });
-
-    if (!confirm) {
-      return;
-    }
-
-    try {
-      const extension = installed.find((x) => x.id === element.uuid);
-      const state = await extensionToFormState(extension);
-      dispatch(actions.resetInstalled(state));
-    } catch (error) {
-      reportError(error);
-      dispatch(actions.adapterError({ uuid: element.uuid, error }));
-    }
-  }, [showConfirmation, dispatch, element.uuid, installed]);
-}
-
-export function useRemove(element: FormState): () => void {
+function useRemove(element: FormState): () => void {
   const { port } = useContext(DevToolsContext);
   const { addToast } = useToasts();
   const { values } = useFormikContext<FormState>();
@@ -128,25 +89,4 @@ export function useRemove(element: FormState): () => void {
   }, [showConfirmation, values, addToast, port, element, dispatch]);
 }
 
-export function useEditable(): Set<string> {
-  const { knownEditable } = useSelector(selectEditor);
-
-  const [initialEditable] = useAsyncState(async () => {
-    const { data } = await axios.get<EditablePackage[]>(
-      await makeURL("api/bricks/"),
-      {
-        headers: { Authorization: `Token ${await getExtensionToken()}` },
-      }
-    );
-    return new Set(data.map((x) => x.name));
-  }, []);
-
-  return useMemo<Set<string>>(() => {
-    // set union
-    const rv = new Set<string>(initialEditable ?? new Set());
-    for (const x of knownEditable) {
-      rv.add(x);
-    }
-    return rv;
-  }, [initialEditable, knownEditable]);
-}
+export default useRemove;
